adds new files to map_validation folder
